# Dawn-News-API/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API that fetches the latest news articles from Dawn.com",
        "title": "Dawn News Web Scraper API",
        "contact": {
            "name": "Abdul Ali Khan"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/article/{id}": {
            "get": {
                "description": "Return a news article by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Return a news article by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                }
            }
        },
        "/latest-news/": {
            "get": {
                "description": "List all latest news articles published on Dawn.com",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Return a list of latest news articles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ArticleDetails"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Article": {
            "type": "object",
            "properties": {
                "story": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.ArticleDetails": {
            "type": "object",
            "properties": {
                "excerpt": {
                    "type": "string"
                },
                "headline": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}